LayoutTests/imported/blink/fast/dom/HTMLImageElement/image-src-onerror.html

<!DOCTYPE html>
<script>
    if (testRunner) {
        testRunner.dumpAsText();
        testRunner.waitUntilDone();
    }
    window.onload = function() {
        setTimeout(function() {
            if (testRunner)
                testRunner.notifyDone();
        }, 500);
    };
</script>
<img src="1x.gif" onerror="console.log('Error triggerred');">
